requirements.txt
requests
pyOpenSSL>=19.1.0